Compress internal SRL binaries...
* Compressed internal SRL binaries for built in
stage2/udisk/nrio-usb-disk. This saves space in the main NDS file and
managed to get main NDS file below 1MB.
* uDisk updator arm9 BIN files updated to flash 8 0x20000 blocks instead
of 9 now that xulumenu is small enough to fit in 8.
* tonccpy moved to ITCM.
* nrio_detect moved to ITCM.
* Moved ram location second bootloader uses as tmp memory to load built
in SRL binaries. Was 0x02200000 but now is 0x02300000. This is to make
room for buffer needed for decompression function.